requirements.txt
google-cloud-bigquery
pandas
pandas-gbq
requests
pyarrow
apache-airflow
python-dotenv
os